+++ random.h
@@ -2,6 +2,7 @@
 #define ZZGO_RANDOM_H
 
 void fast_srandom(unsigned long seed);
-unsigned long fast_random(unsigned long max);
+/* Note that only 16bit numbers can be returned. */
+unsigned long fast_random(unsigned int max);
 
 #endif
